Added support for "isRoutingKey"
Allows support of the cassandra-driver load-balancer `TokenAwarePolicy` (the default load balancing policy) ability to choose the node that the record(s) belong to as the coordinator for a query, based on the given partition keys.

Setting `isRoutingKey` to `true` when using the `helenus` driver, or if not using `TokenAwarePolicy` will cause no effect.